refactor: improve code readability and testing
